requirements.txt
fastapi
uvicorn
pandas
openpyxl
scikit-learn
sentence-transformers
pymongo
python-dotenv
requests
python-multipart
rank_bm25
langchain